Add files via upload
E-commerce called MarketPlace which I developed to understand and learn about back end functionality of websites. This project is similar to JSC's Flask course (https://www.youtube.com/watch?v=Qr4QMBUPxWo&t=13691s). I made very minor UI changes in this project. I intend to add some additional features in the future such as user profile accessibility and so on. My aim is to keep on working on this project to make my Flask, backend, bootstrap and HTML knowledge stronger. Thanks to JSC for this project and the course which has helped me a lot.